+++ tests/testthat/test-distributions.R
@@ -153,6 +153,7 @@ test_succeeds("VonMises distribution works", {
 
 test_succeeds("VectorSinhArcsinhDiag distribution works", {
 
+  skip_if_tfp_above("0.9")
   n <- 10
   scale_diag <- runif(n)
   scale_identity_multiplier <- 1
@@ -174,6 +175,7 @@ test_succeeds("VectorSinhArcsinhDiag distribution works", {
 
 test_succeeds("VectorLaplaceLinearOperator distribution works", {
 
+  skip_if_tfp_above("0.9")
   mu <- c(1, 2, 3)
   cov <-
     matrix(
@@ -192,6 +194,7 @@ test_succeeds("VectorLaplaceLinearOperator distribution works", {
 
 test_succeeds("VectorLaplaceDiag distribution works", {
 
+  skip_if_tfp_above("0.9")
   d <- tfd_vector_laplace_diag(loc = matrix(rep(0, 6), ncol =3))
   expect_equivalent(d %>% tfd_stddev() %>% tensor_value(), rep(sqrt(2), 3), tol = 1e-8)
 })
@@ -1019,33 +1022,36 @@ test_succeeds("WishartTriL distribution works", {
 
 test_succeeds("PixelCNN distribution works", {
 
-  skip_if_tfp_below("0.9")
+  skip_if_tfp_below("0.10")
 
-  # library(tfds)
-  # library(tfdatasets)
+  library(tfdatasets)
   library(keras)
 
-  # data <- tfds_load('mnist')
-  # train_data <- data$train
-  # test_data <- data$test
-  #
-  # image_preprocess <- function(x) {
-  #   x$image <- tf$cast(x$image, tf$float32)
-  #   list(tuple(x$image, x$label))
-  # }
-  #
-  # batch_size <- 16
-  # train_it <- train_data %>% dataset_map(image_preprocess) %>%
-  #   dataset_batch(batch_size) %>%
-  #   dataset_shuffle(1000)
+  mnist <- dataset_mnist()
+  train_data <- mnist$train
+
+  preprocess <- function(record) {
+    record$x <- tf$cast(record$x, tf$float32) %>%
+      tf$expand_dims(axis = -1L)
+    record$y <- tf$cast(record$y, tf$float32)
+    list(tuple(record$x, record$y))
+  }
+
+  batch_size <- 32
+
+  train_ds <- tensor_slices_dataset(train_data)
+  train_ds <- train_ds %>%
+    dataset_take(32) %>%
+    dataset_map(preprocess) %>%
+    dataset_batch(batch_size)
 
   dist <- tfd_pixel_cnn(
     image_shape = c(28, 28, 1),
     conditional_shape = list(),
     num_resnet = 1,
-    num_hierarchies = 2,
-    num_filters = 32,
-    num_logistic_mix = 5,
+    num_hierarchies = 1,
+    num_filters = 2,
+    num_logistic_mix = 1,
     dropout_p = .3
   )
 
@@ -1055,16 +1061,23 @@ test_succeeds("PixelCNN distribution works", {
 
   model <- keras_model(inputs = list(image_input, label_input), outputs = log_prob)
   model$add_loss(-tf$reduce_mean(log_prob))
-  model$compile(
-    optimizer=optimizer_adam(lr = .001),
-    metrics=list())
+  model$compile(optimizer=optimizer_adam(lr = .001))
 
-  # model %>% fit(train_it, epochs = 1)
-  # samples <- dist %>% tfd_sample(1, conditional_input = 1)
-  #
-  # img <- samples[1, , , 1]
-  # img <- t(apply(img, 2, rev))
-  # image(1:28, 1:28, img, col = gray((0:255)/255), xaxt = 'n', yaxt = 'n')
+  model %>% fit(train_ds, epochs = 1)
 
+  samples <- dist %>% tfd_sample(1, conditional_input = 1)
+  expect_equal(dim(samples), c(1, 28, 28, 1))
+
+})
+
+test_succeeds("BetaBinomial distribution works", {
+
+  skip_if_tfp_below("0.10")
+
+  d <- tfd_beta_binomial(
+    total_count = c(5, 10, 20),
+    concentration1 = c(.5, 2, 3),
+    concentration0 = c(4, 3, 2))
+  expect_equal(d %>% tfd_log_prob(1) %>% tensor_value() %>% length(), 3)
 })